pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "maxima-vit"
version = "0.1.0"
description = "Vision Transformer-based 6D detector geometry calibration for 2D XRD patterns"
readme = "README.md"
requires-python = ">=3.11"
license = { file = "LICENSE" }
dependencies = [
    "fabio==2024.9.0",
    "h5py==3.14.0",
    "matplotlib==3.10.6",
    "numpy==2.3.2",
    "pyFAI==2026.2.1",
    "PyYAML==6.0.2",
    "scikit-image>=0.24.0",
    "scikit-learn==1.7.2",
    "scipy>=1.14.0",
    "seaborn==0.13.2",
    "tensorboard>=2.18.0",
    "torch==2.8.0",
    "torchvision==0.23.0",
    "tqdm==4.67.1",
    "transformers==4.56.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-cov>=5.0.0",
]

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["."]
include = ["maxima_vit", "maxima_vit.*", "src", "src.*"]

[tool.pytest.ini_options]
minversion = "8.0"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "unit: Fast unit tests",
    "integration: Multi-module integration tests",
    "slow: Slow tests excluded from default run",
    "script_smoke: Script-level smoke tests",
    "data: Data pipeline tests",
    "model: Model construction and inference tests",
]
addopts = "--strict-markers -q"
norecursedirs = ["lib64"]
